REFACTOR-FE:refactor app.jsx into multiple files {IA}
