+++ client/app/LandingPage/PlanVacationButton.jsx
@@ -1,0 +1,21 @@
+import React, { Component, PropTypes } from 'react';
+import {FancyBorder} from '../helpers.js';
+
+
+const PlanVacationButton = props => (
+  <FancyBorder color="yellow">
+    <button
+      className="centered"
+      onClick={() => props.handlePlanButtonClick()}
+    >
+    Plan My Vacation
+    </button>
+  </FancyBorder>
+
+);
+
+PlanVacationButton.propTypes = {
+  handlePlanButtonClick: PropTypes.func,
+};
+
+export default PlanVacationButton;
